Cheats added but because of highscore adding

+++ STM32/FirstIntegration/header/main.h
@@ -49,7 +49,7 @@ struct Bullet
 struct Player
 {
     uint8_t wave;
-    uint8_t health;
+    uint16_t health;
     uint16_t money;
 };
